% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello. This package is first try to make a package}
\usage{
hello()
}
\arguments{
\item{y}{is the response variable as col in dataframe}

\item{x1}{is factor that needs to group the response together with}

\item{x2}{is the same}
}
\value{
Gives sumtable of dafnia grouped by lake and temperature
}
\description{
The function returns a sumtable and made for the dataset dafnia
}
\examples{
sumtable(dafnia$Feeding,dafnia$Temp,dafnia$Lake) - if you put in a default then write it here
}
